requirements.txt
uvicorn==0.13.3
fastapi==0.63.0
SQLAlchemy==1.3.22
PyMySQL==0.10.1
PyYAML==5.3.1